+++ .gitignore
@@ -1,0 +1,13 @@
+.idea
+
+# Binaries for programs and plugins
+*.exe
+*.exe~
+*.dll
+*.so
+*.dylib
+*.test
+
+*.out
+*.log
+*.csv
