+++ go.mod
@@ -1,0 +1,3 @@
+module wayback-keyword-search
+
+go 1.21.6
